// src/app/pages/model/administrator/systemAdmin/working.day.model.ts
export interface WorkingDayModel {
    id?: number;
    working_day_name: string;
    is_monday: boolean;
    is_tuesday: boolean;
    is_wednesday: boolean;
    is_thursday: boolean;
    is_friday: boolean;
    is_saturday: boolean;
    is_sunday: boolean;
    b_status: boolean;
    description: string;
}